requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.1
Pillow==9.1.1
python-docx==0.8.11
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
